moved indexes and transformers out of audit instance into the raven persistence project
