+++ src/ServiceControl.Audit.Persistence.RavenDb/Indexes/ExpiryKnownEndpointsIndex.cs
@@ -1,4 +1,4 @@
-﻿namespace ServiceControl.Audit.Persistence.RavenDB.Expiration
+﻿namespace ServiceControl.Audit.Persistence.RavenDb.Indexes
 {
     using System.Linq;
     using Monitoring;
