+++ src/ServiceControl.Audit.UnitTests/ApprovalFiles/APIApprovals.PublicClr.approved.txt
@@ -15,10 +15,6 @@ namespace ServiceControl.Audit.Auditing
         public System.Guid Id { get; set; }
         public ServiceControl.Audit.Auditing.FailedTransportMessage Message { get; set; }
     }
-    public class FailedAuditImportIndex : Raven.Client.Indexes.AbstractIndexCreationTask<ServiceControl.Audit.Auditing.FailedAuditImport>
-    {
-        public FailedAuditImportIndex() { }
-    }
     public class FailedTransportMessage
     {
         public FailedTransportMessage() { }
@@ -63,39 +59,6 @@ namespace ServiceControl.Audit.Auditing.MessagesView
         public ServiceControl.Audit.Monitoring.EndpointDetails SendingEndpoint { get; set; }
         public ServiceControl.Audit.Monitoring.MessageStatus Status { get; set; }
         public System.DateTime? TimeSent { get; set; }
-    }
-    public class MessagesViewIndex : Raven.Client.Indexes.AbstractIndexCreationTask<ServiceControl.Audit.Auditing.ProcessedMessage, ServiceControl.Audit.Auditing.MessagesView.MessagesViewIndex.SortAndFilterOptions>
-    {
-        public MessagesViewIndex() { }
-        public class SortAndFilterOptions
-        {
-            public SortAndFilterOptions() { }
-            public string ConversationId { get; set; }
-            public System.TimeSpan? CriticalTime { get; set; }
-            public System.TimeSpan? DeliveryTime { get; set; }
-            public bool IsSystemMessage { get; set; }
-            public string MessageId { get; set; }
-            public string MessageType { get; set; }
-            public System.DateTime ProcessedAt { get; set; }
-            public System.TimeSpan? ProcessingTime { get; set; }
-            public string[] Query { get; set; }
-            public string ReceivingEndpointName { get; set; }
-            public ServiceControl.Audit.Monitoring.MessageStatus Status { get; set; }
-            public System.DateTime TimeSent { get; set; }
-        }
-    }
-    public class MessagesViewTransformer : Raven.Client.Indexes.AbstractTransformerCreationTask<ServiceControl.Audit.Auditing.MessagesView.MessagesViewTransformer.Result>
-    {
-        public MessagesViewTransformer() { }
-        public class Result
-        {
-            public Result() { }
-            public System.Collections.Generic.Dictionary<string, string> Headers { get; set; }
-            public string Id { get; set; }
-            public System.Collections.Generic.Dictionary<string, object> MessageMetadata { get; set; }
-            public System.DateTime ProcessedAt { get; set; }
-            public string UniqueMessageId { get; set; }
-        }
     }
 }
 namespace ServiceControl.Audit.Infrastructure.Settings
