+++ src/ServiceControl.Audit.UnitTests/RavenPersistedTypes.cs
@@ -2,7 +2,6 @@
 {
     using System;
     using System.Linq;
-    using Audit.Infrastructure;
     using NUnit.Framework;
     using Particular.Approvals;
     using Raven.Client.Indexes;
@@ -13,11 +12,6 @@
         [Test]
         public void Verify()
         {
-            var documentTypes = typeof(Bootstrapper).Assembly.GetTypes()
-                .Where(type => typeof(AbstractIndexCreationTask).IsAssignableFrom(type))
-                .SelectMany(indexType => indexType.BaseType?.GenericTypeArguments)
-                .Distinct();
-
             var ravenPersistenceType = Type.GetType(DataStoreConfig.RavenDbPersistenceTypeFullyQualifiedName, true);
 
             var ravenPersistenceTypes = ravenPersistenceType.Assembly.GetTypes()
@@ -25,7 +19,7 @@
                 .SelectMany(indexType => indexType.BaseType?.GenericTypeArguments)
                 .Distinct();
 
-            var documentTypeNames = string.Join(Environment.NewLine, ravenPersistenceTypes.Select(t => t.AssemblyQualifiedName).Union(documentTypes.Select(t => t.AssemblyQualifiedName)));
+            var documentTypeNames = string.Join(Environment.NewLine, ravenPersistenceTypes.Select(t => t.AssemblyQualifiedName));
 
             Approver.Verify(documentTypeNames);
         }
